@@ -11,8 +11,8 @@
         "eslint-plugin-prettier": "^3.4.1",
         "lodash": "^4.17.15",
         "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-router-dom": "^5.3.0",
         "react-scripts": "4.0.3",
         "react-select": "^5.1.0",
