@@ -1,18 +1,18 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import "bootstrap/dist/css/bootstrap.css";
 import App from "./app/App";
 import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(
-    <React.StrictMode>
-        <BrowserRouter>
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+    <BrowserRouter>
+        <React.StrictMode>
             <App />
-        </BrowserRouter>
-    </React.StrictMode>,
-    document.getElementById("root")
+        </React.StrictMode>
+    </BrowserRouter>
 );
 
 // If you want to start measuring performance in your app, pass a function
