@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
 
-const TextField = ({ label, type, name, value, onChange, error }) => {
+const TextField = ({ label, type, name, value, onChange, error, ...rest }) => {
     const [showPassword, setShowPassword] = useState(false);
 
     const handleChange = ({ target }) => {
@@ -24,6 +24,7 @@ const TextField = ({ label, type, name, value, onChange, error }) => {
                     value={value}
                     onChange={handleChange}
                     className={getInputClasses()}
+                    {...rest}
                 />
 
                 {type === "password" && (
@@ -56,4 +57,4 @@ TextField.propTypes = {
     error: PropTypes.string
 };
 
-export default TextField;
+export default React.memo(TextField);
