@@ -7,33 +7,31 @@ import Divider from "../../common/divider";
 import withLogin from "./withLogin";
 import withPropsStyles from "./withPropsStyles";
 const HOCExample = () => {
-    const ComponentWithAuth = withLogin(Component);
-    const ComponentWithPropsStyles = withPropsStyles(Component);
-    const NewComponent = withPropsStyles(ComponentWithAuth);
-    return (
-        <>
-            <CardWrapper>
-                <SmallTitle>1. Обычный компонент</SmallTitle>
-                <Divider />
-                <Component />
-            </CardWrapper>
-            <CardWrapper>
-                <SmallTitle>2. Функциональный HOC</SmallTitle>
-                <Divider />
-                <ComponentWithAuth />
-            </CardWrapper>
-            <CardWrapper>
-                <SmallTitle>3. HOC With Styles and Props</SmallTitle>
-                <Divider />
-                <ComponentWithPropsStyles />
-            </CardWrapper>
-            <CardWrapper>
-                <SmallTitle>4. Composed HOC</SmallTitle>
-                <Divider />
-                <NewComponent />
-            </CardWrapper>
-        </>
-    );
+  const ComponentWithAuth = withLogin(Component);
+  const ComponentWithPropsStyles = withPropsStyles(Component);
+  const NewComponent = withPropsStyles(ComponentWithAuth);
+  return (
+    <>
+      <CardWrapper>
+        <SmallTitle>1. Обычный компонент</SmallTitle>
+        <Divider />
+        <Component />
+      </CardWrapper>
+      <CardWrapper>
+        <SmallTitle>2. Функциональный HOC</SmallTitle>
+        <Divider />
+        <ComponentWithAuth />
+      </CardWrapper>
+      <CardWrapper>
+        <SmallTitle>3. HOC With Styles and Props</SmallTitle>
+        <ComponentWithPropsStyles />
+      </CardWrapper>
+      <CardWrapper>
+        <SmallTitle>4. Composed HOC</SmallTitle>
+        <NewComponent />
+      </CardWrapper>
+    </>
+  );
 };
 
 export default HOCExample;
