@@ -2,12 +2,12 @@ import React from "react";
 import SmallTitle from "../../common/typografy/smallTitle";
 
 const withLogin = (Component) => (props) => {
-    const isLogin = localStorage.getItem("auth");
-    return (
-        <>
-            {isLogin ? <Component {...props} /> : <SmallTitle>Auth</SmallTitle>}
-        </>
-    );
+  const isLogin = localStorage.getItem("auth");
+  return (
+    <>
+      {isLogin ? <Component {...props} /> : <SmallTitle>Auth</SmallTitle>}
+    </>
+  );
 };
 
 export default withLogin;
