@@ -1,56 +1,55 @@
-import React, { useState, useRef, useEffect, useCallback } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import CardWrapper from "../../common/Card";
 import SmallTitle from "../../common/typografy/smallTitle";
 
 const UseCallBackExample = () => {
-    const [data, setDate] = useState({});
-    const withOutCallback = useRef(0);
-    const withCallback = useRef(0);
+  const [data, setData] = useState({});
+  const withOutCallback = useRef(0);
+  const withCallback = useRef(0);
 
-    const handleChange = ({ target }) => {
-        setDate((prevState) => ({ ...prevState, [target.name]: target.value }));
-    };
-    // WuthOut Callback
-    const validateWithOutCallback = (data) => {
-        console.log(data);
-    };
+  const handleChange = ({ target }) => {
+    setData((prevState) => ({ ...prevState, [target.name]: target.value }));
+  };
 
-    useEffect(() => {
-        withOutCallback.current++;
-    }, [validateWithOutCallback]);
+  // Without Callback
+  const validateWithOutCallback = (data) => {
+    console.log(data);
+  };
+  useEffect(() => {
+    withOutCallback.current++;
+  }, [validateWithOutCallback]);
 
-    // With Callback
-    const validateWithCallback = useCallback((data) => {
-        console.log(data);
-    }, []);
-    useEffect(() => {
-        withCallback.current++;
-    }, [validateWithCallback]);
+  // With Callback
+  const validateWithCallback = useCallback((data) => {
+    console.log(data);
+  }, []);
+  useEffect(() => {
+    withCallback.current++;
+  }, [validateWithCallback]);
 
-    useEffect(() => {
-        validateWithOutCallback(data);
-        validateWithCallback(data);
-    }, [data]);
+  useEffect(() => {
+    validateWithOutCallback(data);
+    validateWithCallback(data);
+  }, [data]);
 
-    return (
-        <CardWrapper>
-            <SmallTitle>Example</SmallTitle>
-            <p>Render withOutCallback: {withOutCallback.current}</p>
-            <p>Render withCallback: {withCallback.current}</p>
-
-            <label htmlFor="email" className="form-label">
-                Email
-            </label>
-            <input
-                type="email"
-                className="form-control mb-4"
-                id="email"
-                value={data.email || ""}
-                onChange={handleChange}
-                name="email"
-            />
-        </CardWrapper>
-    );
+  return (
+    <CardWrapper>
+      <SmallTitle>Example</SmallTitle>
+      <p> Render withOutCallback: {withOutCallback.current}</p>
+      <p> Render withCallback: {withCallback.current}</p>
+      <label htmlFor="email" className="form-label">
+        Email
+      </label>
+      <input
+        type="email"
+        className="form-control"
+        id="email"
+        value={data.email || ""}
+        name="email"
+        onChange={handleChange}
+      />
+    </CardWrapper>
+  );
 };
 
 export default UseCallBackExample;
