@@ -1,18 +1,13 @@
-import React, { useRef, useEffect } from "react";
+import React, { useRef } from "react";
 import CollapseWrapper from "../common/collapse";
 const UseRefExercise = () => {
-    const prevState = useRef();
+    const blockRef = useRef();
 
-    useEffect(() => {
-        console.log(prevState);
-    });
-    const toggleBlockText = () => {
-        prevState.current.style.width = "150px";
-        prevState.current.style.height = "80px";
-        prevState.current.innerText = "Text";
-        console.log(prevState.current);
+    const handleClick = () => {
+        blockRef.current.children[0].innerText = "text";
+        blockRef.current.style.width = "150px";
+        blockRef.current.style.height = "80px";
     };
-
     return (
         <CollapseWrapper title="Упражнение">
             <p className="mt-3">
@@ -24,7 +19,7 @@ const UseRefExercise = () => {
                 <li>высота и ширина станут равны 150 и 80 соответственно</li>
             </ul>
             <div
-                ref={prevState}
+                ref={blockRef}
                 className="bg-primary d-flex flex-row justify-content-center align-items-center rounded"
                 style={{
                     height: 40,
@@ -34,8 +29,8 @@ const UseRefExercise = () => {
             >
                 <small>Блок</small>
             </div>
-            <button className="btn btn-primary mt-4" onClick={toggleBlockText}>
-                Submit
+            <button className="btn btn-secondary my-4" onClick={handleClick}>
+                Trigger action
             </button>
         </CollapseWrapper>
     );
