@@ -1,9 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useHistory, useParams } from "react-router-dom";
-import { validator } from "../../../utils/validator";
 import api from "../../../api";
-import TextField from "../../common/form/textField";
-import SelectField from "../../common/form/selectField";
+import FormComponent, { TextField, SelectField } from "../../common/form";
 
 const EditUserPage = () => {
     const { userId } = useParams();
@@ -15,53 +13,32 @@ const EditUserPage = () => {
         profession: ""
     });
     const [professions, setProfession] = useState([]);
-    const [qualities, setQualities] = useState({});
-    const [errors, setErrors] = useState({});
+
     const getProfessionById = (id) => {
         for (const prof in professions) {
             const profData = professions[prof];
             if (profData._id === id) return profData;
         }
     };
-    const getQualities = (elements) => {
-        const qualitiesArray = [];
-        for (const elem of elements) {
-            for (const quality in qualities) {
-                if (elem.value === qualities[quality]._id) {
-                    qualitiesArray.push(qualities[quality]);
-                }
-            }
-        }
-        return qualitiesArray;
-    };
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        const isValid = validate();
-        if (!isValid) return;
-        const { profession, qualities } = data;
+
+    const handleSubmit = (data) => {
+        const { profession } = data;
         api.users
-          .update(userId, {
-              ...data,
-              profession: getProfessionById(profession),
-              qualities: getQualities(qualities)
-          })
-          .then((data) => history.push(`/users/${data._id}`));
-        console.log(data);
-    };
-    const transformData = (data) => {
-        return data.map((qual) => ({ label: qual.name, value: qual._id }));
+            .update(userId, {
+                ...data,
+                profession: getProfessionById(profession)
+            })
+            .then((data) => history.push(`/users/${data._id}`));
     };
     useEffect(() => {
         setIsLoading(true);
-        api.users.getById(userId).then(({ profession, qualities, ...data }) =>
-          setData((prevState) => ({
-              ...prevState,
-              ...data,
-              qualities: transformData(qualities),
-              profession: profession._id
-          }))
+        api.users.getById(userId).then(({ profession, ...data }) =>
+            setData((prevState) => ({
+                ...prevState,
+                ...data,
+                profession: profession._id
+            }))
         );
-        api.qualities.fetchAll().then((data) => setQualities(data));
         api.professions.fetchAll().then((data) => setProfession(data));
     }, []);
     useEffect(() => {
@@ -83,62 +60,39 @@ const EditUserPage = () => {
             }
         }
     };
-    useEffect(() => validate(), [data]);
-    const handleChange = (target) => {
-        setData((prevState) => ({
-            ...prevState,
-            [target.name]: target.value
-        }));
-    };
-    const validate = () => {
-        const errors = validator(data, validatorConfig);
-        setErrors(errors);
-        return Object.keys(errors).length === 0;
-    };
-    const isValid = Object.keys(errors).length === 0;
+
     return (
-      <div className="container mt-5">
-          <div className="row">
-              <div className="col-md-6 offset-md-3 shadow p-4">
-                  {!isLoading && Object.keys(professions).length > 0 ? (
-                    <form onSubmit={handleSubmit}>
-                        <TextField
-                          label="Имя"
-                          name="name"
-                          value={data.name}
-                          onChange={handleChange}
-                          error={errors.name}
-                        />
-                        <TextField
-                          label="Электронная почта"
-                          name="email"
-                          value={data.email}
-                          onChange={handleChange}
-                          error={errors.email}
-                        />
-                        <SelectField
-                          label="Выбери свою профессию"
-                          defaultOption="Choose..."
-                          options={professions}
-                          name="profession"
-                          onChange={handleChange}
-                          value={data.profession}
-                          error={errors.profession}
-                        />
-                        <button
-                          type="submit"
-                          disabled={!isValid}
-                          className="btn btn-primary w-100 mx-auto"
+        <div className="container mt-5">
+            <div className="row">
+                <div className="col-md-6 offset-md-3 shadow p-4">
+                    {!isLoading && Object.keys(professions).length > 0 ? (
+                        <FormComponent
+                            onSubmit={handleSubmit}
+                            validatorConfig={validatorConfig}
+                            defaultData={data}
                         >
-                            Обновить
-                        </button>
-                    </form>
-                  ) : (
-                    "Loading..."
-                  )}
-              </div>
-          </div>
-      </div>
+                            <TextField label="Имя" name="name" autoFocus />
+                            <TextField label="Электронная почта" name="email" />
+                            <SelectField
+                                label="Выбери свою профессию"
+                                defaultOption="Choose..."
+                                options={professions}
+                                name="profession"
+                            />
+
+                            <button
+                                type="submit"
+                                className="btn btn-primary w-100 mx-auto"
+                            >
+                                Обновить
+                            </button>
+                        </FormComponent>
+                    ) : (
+                        "Loading..."
+                    )}
+                </div>
+            </div>
+        </div>
     );
 };
 
